refactor: moves deployment module
