+++ cli/cmd/cmds/module/dump.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	"github.com/input-output-hk/catalyst-forge/cli/pkg/run"
-	"github.com/input-output-hk/catalyst-forge/lib/project/deployment"
+	"github.com/input-output-hk/catalyst-forge/lib/deployment"
 	"github.com/input-output-hk/catalyst-forge/lib/tools/fs"
 )
