+++ foundry/operator/internal/controller/suite_env_test.go
@@ -13,6 +13,11 @@ import (
 	"github.com/go-git/go-git/v5/plumbing/object"
 	"github.com/go-git/go-git/v5/storage"
 	"github.com/go-git/go-git/v5/storage/filesystem"
+	foundryv1alpha1 "github.com/input-output-hk/catalyst-forge/foundry/operator/api/v1alpha1"
+	"github.com/input-output-hk/catalyst-forge/foundry/operator/pkg/config"
+	"github.com/input-output-hk/catalyst-forge/foundry/operator/pkg/handlers"
+	"github.com/input-output-hk/catalyst-forge/lib/deployment"
+	"github.com/input-output-hk/catalyst-forge/lib/deployment/deployer"
 	"github.com/input-output-hk/catalyst-forge/lib/foundry/client"
 	"github.com/input-output-hk/catalyst-forge/lib/foundry/client/auth"
 	"github.com/input-output-hk/catalyst-forge/lib/foundry/client/deployments"
@@ -20,11 +25,6 @@ import (
 	"github.com/input-output-hk/catalyst-forge/lib/foundry/client/github"
 	"github.com/input-output-hk/catalyst-forge/lib/foundry/client/releases"
 	"github.com/input-output-hk/catalyst-forge/lib/foundry/client/users"
-	foundryv1alpha1 "github.com/input-output-hk/catalyst-forge/foundry/operator/api/v1alpha1"
-	"github.com/input-output-hk/catalyst-forge/foundry/operator/pkg/config"
-	"github.com/input-output-hk/catalyst-forge/foundry/operator/pkg/handlers"
-	"github.com/input-output-hk/catalyst-forge/lib/project/deployment"
-	"github.com/input-output-hk/catalyst-forge/lib/project/deployment/deployer"
 	tu "github.com/input-output-hk/catalyst-forge/lib/project/utils/test"
 	"github.com/input-output-hk/catalyst-forge/lib/providers/secrets"
 	sb "github.com/input-output-hk/catalyst-forge/lib/schema/blueprint"
