+++ lib/deployment/manifest_gen.go
