+++ lib/deployment/mocks/manifest.go
