+++ lib/deployment/providers/helm/generator_test.go
@@ -11,7 +11,7 @@ import (
 
 	"cuelang.org/go/cue"
 	"cuelang.org/go/cue/cuecontext"
-	"github.com/input-output-hk/catalyst-forge/lib/project/deployment/providers/helm/downloader/mocks"
+	"github.com/input-output-hk/catalyst-forge/lib/deployment/providers/helm/downloader/mocks"
 	sp "github.com/input-output-hk/catalyst-forge/lib/schema/blueprint/project"
 	"github.com/input-output-hk/catalyst-forge/lib/tools/testutils"
 	"github.com/stretchr/testify/require"
