+++ lib/deployment/providers/helm/testdata/golden.yaml
