+++ lib/deployment/providers/helm/testdata/templates/deployment.yaml
