+++ lib/deployment/providers/kcl/generator_test.go
@@ -6,8 +6,8 @@ import (
 
 	"cuelang.org/go/cue"
 	"cuelang.org/go/cue/cuecontext"
-	"github.com/input-output-hk/catalyst-forge/lib/project/deployment/providers/kcl/client"
-	"github.com/input-output-hk/catalyst-forge/lib/project/deployment/providers/kcl/client/mocks"
+	"github.com/input-output-hk/catalyst-forge/lib/deployment/providers/kcl/client"
+	"github.com/input-output-hk/catalyst-forge/lib/deployment/providers/kcl/client/mocks"
 	sp "github.com/input-output-hk/catalyst-forge/lib/schema/blueprint/project"
 	"github.com/input-output-hk/catalyst-forge/lib/tools/fs/billy"
 	"github.com/input-output-hk/catalyst-forge/lib/tools/testutils"
